Feature:   1. Support public cloud shared proxy connection number resource      isolation   2. Support public cloud shared proxy tenant level whitelist   3. Support shared proxy cluster configuration push   4. Support hot restart   5. Supports partition routing for Timestmp data types   6. Support client-side Execute optimization, allow without Param Type   7. Proxyd.sh supports port modification, rslist startup, and configurl   startup   8. PS memory optimization   9. 16C model performance optimization   10. When accessing the cluster for the first time, after OBProxy   obtains the RSList from the ConfigServer, the All Server List is   obtained asynchronously and adjusted to obtain it synchronously   11. Log compression can be turned on with   enable_syslog_file_compress=True   12. Full username strict mode can be enabled by   enable_full_username=True   13. You can increase the SQL length of obproxy_digest.log by setting   the digest_sql_length parameter   14. Added support for partition routing for timestamp type and   negative numbers   15. Added support for DDL statements in sharding scenarios   16. Added tenant-level SSL link encryption   17. Supports splicing URLs by zone   18. OB Sharding supports Join
Bugfix:
  1. Fix the problem that the new password under the newly deployed
  cluster etc is inconsistent with the old password
  2. Fix ODP unable to handle configurations larger than 4K, currently
  there is no limit to configuration size
  3. Fix the cluster without LDG. After the enable_ldg switch is turned
  on, the proxy stops updating the location cache, resulting in no sense
  of offline machines.
  4. Fix Under the 2.0 protocol, the data returned to the driver by the ps
  and fetch protocols is incorrect, resulting in the client hang problem
  5. Fix mutli swichover obproxy core at
  databuff_print_key_obj<oceanbase::common::ObLogPrintPointerCntx
  6. Fix obproxy memory leak caused by frequent push of current
  limiting rules
  7. Fix inaccurate routing in col = ("123") mode
  8. Fix too many current-limiting rules causing the push to fail to
  take effect
  9. Fix obproxy memory leak caused by frequent push of current
  limiting rules
  10. Fix After repairing the failover of the standby database into the main
  database, after the original main database is down, the proxy will
  have a short time to route the request to the original main database
  11. Fix Failed to establish connection without specifying cluster id
  within 3 minutes after lossless FO
  12. Fix After repairing the deletion of the standby database, continue to
  send requests to the old clusterId
  13. Fix the problem that the new protocol link proxy cursor scene,
  when the fench reaches the second value, it will be rammed
  14. Fix the problem of ps cursor ramming under non-compression
  protocol
  15. Fix a new connection reporting too many connections error
  16. Fix the case-insensitive name field of the ssl_config
  configuration table, resulting in unreasonable configuration reading
  logic
  17. Fix Compression and 2.0 protocol, obdump import data will be hung
  18. Fix the new protocol, refcursor cannot get the value when there
  are other out parameters
  19. Fix obproxy core at oceanbase::common::ObObjCmpFuncs::compare null
  safe
  20. Fix obproxy core at
  oceanbase::common::ObTimeConverter::str_to_digit_with_date
  21. Fix the error -5560, Unknown prepared statement handle when bmsql
  opens ps when the compression protocol is enabled

+++ src/lib/objectpool/ob_concurrency_objpool.cpp
@@ -292,7 +292,7 @@ inline bool ObObjFreeList::need_to_reclaim(ObThreadCache *thread_cache)
 {
   bool ret = false;
   if (reclaim_opt_.enable_reclaim_) {
-    if(thread_cache->nr_free_ >= thread_cache->nr_average_
+    if(thread_cache->nr_free_ >= static_cast<int64_t>(thread_cache->nr_average_)
        && thread_cache->nr_free_chunks_ > chunk_count_
        && thread_cache->nr_total_ > std::max(obj_count_base_, obj_count_per_chunk_)) {
       if (thread_cache->nr_overage_++ >= reclaim_opt_.max_overage_) {
